change constants to maps

+++ src/main/java/frc/robot/subsystems/ArmSubsystem.java
@@ -22,12 +22,12 @@ public class ArmSubsystem extends SubsystemBase {
         ArmMotor2.follow(ArmMotor, true);
     }
 
-    public void changeArmSmartMotionParameters(int maxVel, int maxAccel) {
+    public void changeArmSmartMotionParameters(double maxVel, double maxAccel) {
         ArmMotorPIDController.setSmartMotionMaxVelocity(maxVel, 0);
         ArmMotorPIDController.setSmartMotionMaxAccel(maxAccel, 0);
     }
 
-    public void goToPosition(int position) {
+    public void goToPosition(double position) {
         ArmMotorPIDController.setReference(position, CANSparkMax.ControlType.kSmartMotion);
     }
 }
